+++ app/src/main/java/tech/kekulta/jobfinder/presentation/ui/recycler/adapters/OfferBlockAdapter.kt
@@ -45,7 +45,7 @@ class OfferBlockAdapter :
                 this.adapter = offersAdapter
                 this.layoutManager = offersLayoutManager
                 this.itemAnimator?.changeDuration = 0
-                this.addItemDecoration(RecyclerViewRightMargin(context.dp(8).toInt()))
+                this.addItemDecoration(RecyclerViewRightMargin(context.dp(8)))
             }
         }
